requirements.txt
langchain-google-genai
flask
python-dotenv